requirements.txt
playwright
python-dotenv
typer[all]
